+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 * Suppress addition of superfluous parentheses when converting operator matcher that have argument in parentheses to non-operator matcher (e.g. from `== (2 - 1)` to `eq(2 - 1)`)
+* Support auto-modification of syntax configuration in `RSpec.configure`
 
 ## v0.0.2